CRUD schedule + filter booking theo dayStart

+++ src/services/ScheduleService.js
@@ -2,15 +2,47 @@ import Schedule from '../models/Schedule.js'
 import RoomType from '../models/RoomType.js'
 import Room from '../models/Room.js'
 import Hotel from '../models/Hotel.js'
+import Booking from '../models/Booking.js'
+import jwt from 'jsonwebtoken'
 
 const createSchedule = (schedule) => {
     return new Promise(async (resolve, reject) => {
         try {
+            const scheduleDayStart = new Date(schedule.dayStart)
+            const scheduleDayEnd = new Date(schedule.dayEnd)
+            if (!schedule?.bookingId) {
+                return resolve({
+                    status: 'ERR',
+                    message: 'Thiếu bookingId',
+                    statusCode: 400
+                });
+            }
+            const checkBooking = await Booking.findOne({ bookingId: schedule.bookingId })
+            if (checkBooking === null) {
+                return resolve({
+                    status: 'ERR',
+                    message: 'Không tìm thấy đơn đặt phòng tương ứng',
+                    statusCode: 404
+                })
+            }
+            const allScheduleInRoom = await Schedule.find({ roomId: schedule.roomId })
+            const duplicatedSchedule = allScheduleInRoom.filter(sche => (scheduleDayStart <= sche.dayEnd && scheduleDayStart >= sche.dayStart)
+                || (scheduleDayEnd <= sche.dayEnd && scheduleDayEnd >= sche.dayStart))
+            if (duplicatedSchedule.length > 0) {
+                return resolve({
+                    status: 'ERR',
+                    message: 'Không thể tạo lịch vì phòng đã sử dụng trong ngày này',
+                    statusCode: 400
+                })
+            }
             await Schedule.create(schedule)
+            checkBooking.roomQuantity += 1
+            await checkBooking.save()
 
             resolve({
                 status: 'OK',
                 message: 'Tạo lịch đặt phòng thành công',
+                statusCode: 200
             })
 
         } catch (e) {
@@ -22,6 +54,7 @@ const createSchedule = (schedule) => {
 const updateSchedule = (schedule, id) => {
     return new Promise(async (resolve, reject) => {
         try {
+            const newDayEnd = new Date(schedule.dayEnd)
             const checkSchedule = await Schedule.findOne({
                 scheduleId: id
             })
@@ -32,7 +65,27 @@ const updateSchedule = (schedule, id) => {
                     statusCode: 404
                 })
             }
-
+            if (newDayEnd <= checkSchedule.dayEnd) {
+                return resolve({
+                    status: 'ERR',
+                    message: 'dayEnd mới phải lớn hơn dayEnd ban đầu',
+                    statusCode: 400
+                });
+            }
+            const allScheduleInRoom = await Schedule.find({ roomId: checkSchedule.roomId })
+            const duplicatedSchedule = allScheduleInRoom.filter(sche => {
+                if (sche.scheduleId === Number(id)) return false;
+                return (
+                    newDayEnd <= sche.dayEnd && newDayEnd >= sche.dayStart
+                )
+            })
+            if (duplicatedSchedule.length > 0) {
+                return resolve({
+                    status: 'ERR',
+                    message: 'Không thể cập nhật dayEnd mới vì phòng đã sử dụng trong ngày này',
+                    statusCode: 400
+                })
+            }
             await Schedule.findOneAndUpdate({ scheduleId: id },
                 schedule,
                 { new: true })
@@ -61,10 +114,19 @@ const deleteSchedule = (id) => {
                     statusCode: 404
                 })
             }
+            const checkBooking = await Booking.findOne({ bookingId: checkSchedule.bookingId })
+            if (checkBooking.status !== "Đã thanh toán") {
+                return resolve({
+                    status: 'ERR',
+                    message: 'Chỉ có thể xóa lịch hẹn của đơn đặt phòng đã thanh toán',
+                    statusCode: 400
+                })
+            }
             //delete schedule
-
             await Schedule.findOneAndDelete({ scheduleId: id },
                 { new: true })
+            checkBooking.roomQuantity -= 1
+            await checkBooking.save()
 
             resolve({
                 status: 'OK',
@@ -83,7 +145,14 @@ const getDetailSchedule = (id) => {
         try {
             const checkSchedule = await Schedule.findOne({
                 scheduleId: id
-            })
+            }).populate({
+                path: 'bookingId',
+                model: 'Booking',
+                localField: 'bookingId',
+                foreignField: 'bookingId',
+                select: 'bookingCode'
+            }).lean()
+
             if (checkSchedule === null) {
                 return resolve({
                     status: 'ERR',
@@ -105,18 +174,30 @@ const getDetailSchedule = (id) => {
     })
 }
 
-const getAllSchedule = () => {
+const getAllSchedule = (headers) => {
     return new Promise(async (resolve, reject) => {
         try {
-            const checkSchedule = await Schedule.find()
-            if (checkSchedule.length === 0) {
-                return resolve({
-                    status: 'ERR',
-                    message: 'Lịch đặt phòng không tồn tại',
-                    statusCode: 404
-                })
-            }
-
+            const token = headers.authorization.split(' ')[1]
+            const decoded = jwt.verify(token, process.env.ACCESS_TOKEN)
+            const checkHotel = await Hotel.findOne({ userId: decoded.userId, isDeleted: false })
+            const checkRoomType = await RoomType.find({ hotelId: checkHotel.hotelId })
+            const roomTypeIds = checkRoomType.map(roomtype => roomtype.roomTypeId)
+            const checkRoom = await Room.find({ roomTypeId: { $in: roomTypeIds }, isActive: true })
+            const roomIds = checkRoom.map(room => room.roomId)
+            let checkSchedule = await Schedule.find({
+                roomId: { $in: roomIds }
+            }).populate({
+                path: 'bookingId',
+                model: 'Booking',
+                localField: 'bookingId',
+                foreignField: 'bookingId',
+                select: 'bookingCode'
+            }).lean()
+            checkSchedule = checkSchedule.map(schedule => ({
+                ...schedule,
+                bookingCode: schedule.bookingId?.bookingCode,
+                bookingId: schedule.bookingId?.bookingId,
+            }));
             resolve({
                 status: 'OK',
                 message: 'Xem tất cả lịch đặt phòng thành công',
@@ -125,6 +206,7 @@ const getAllSchedule = () => {
             })
 
         } catch (e) {
+            console.log(e)
             reject(e)
         }
     })
